backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
urllib3<2.0
